+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Bug fixes:
 * Ensure native `VALUE`s returned from C are unwrapped before the objects can be collected (@aardvark179).
 * Fix `Enumerator::Lazy#with_index` to start with new index for multiple enumerations (@bjfish).
 * Fix `rb_id2name` to ensure the native string will have the same lifetime as the id (#2630, @aardvark179).
+* Fix `MatchData#[]` exception when passing a length argument larger than the number of match values (#2636, @nirvdrum).
+* Fix `MatchData#[]` exception when supplying a large negative index along with a length argument (@nirvdrum).
 
 Compatibility: